requirements.txt
dimod==0.12.20
dwave-neal==0.6.0
dwave-samplers==1.5.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
sympy==1.13.3